+++ ref/verify.h
@@ -4,7 +4,7 @@
 #include <stddef.h>
 #include <stdint.h>
 
-int PQCLEAN_NAMESPACE_verify(const uint8_t *a, const uint8_t *b, size_t len);
+uint8_t PQCLEAN_NAMESPACE_verify(const uint8_t *a, const uint8_t *b, size_t len);
 
 void PQCLEAN_NAMESPACE_cmov(uint8_t *r, const uint8_t *x, size_t len, uint8_t b);
